Fix: add single quotes to escape pwsh call operator
`&` is treated as a call operator in pwsh. This commit escapes it by adding single quotes `'`.
`&` may also be escaped by using three double quotes `"""` or adding `--%` in front of arguments to prevent pwsh to expand args, though which may not be fully compatible with MS Command or Linux Shell.